Fixed found issues
